create glmalloc module
