+++ pwnlib/heap/glmalloc/arena/arena.py
@@ -1,8 +1,8 @@
-from pwnlib.heap.bins import NoTcacheError
-from pwnlib.heap.malloc_state import \
+from pwnlib.heap.glmalloc.bins import NoTcacheError
+from pwnlib.heap.glmalloc.malloc_state import \
     SMALL_BINS_START_INDEX, \
     LARGE_BINS_START_INDEX
-from pwnlib.heap.basic_formatter import BasicFormatter
+from pwnlib.heap.glmalloc.basic_formatter import BasicFormatter
 
 
 class Arena(object):
