+++ pwnlib/heap/glmalloc/bins/bin_parser.py
@@ -1,6 +1,6 @@
-from pwnlib.heap.bins.small_bin import SmallBins, SmallBin, SmallBinEntry
-from pwnlib.heap.bins.large_bin import LargeBins, LargeBin, LargeBinEntry
-from pwnlib.heap.bins.unsorted_bin import \
+from pwnlib.heap.glmalloc.bins import SmallBins, SmallBin, SmallBinEntry
+from pwnlib.heap.glmalloc.bins import LargeBins, LargeBin, LargeBinEntry
+from pwnlib.heap.glmalloc.bins import \
     UnsortedBins, \
     UnsortedBin, \
     UnsortedBinEntry
