+++ pwnlib/heap/glmalloc/process_informer.py
@@ -1,6 +1,6 @@
 import os.path
 import pwnlib.util.proc
-from pwnlib.heap.utils import \
+from pwnlib.heap.glmalloc.utils import \
     u64, \
     u32, \
     get_libc_version_from_name, \
