+++ pwnlib/heap/glmalloc/utils.py
